Added turn tests Added turn class Modified runner file
