+++ lib/menu.rb
@@ -3,6 +3,7 @@ require 'minitest/pride'
 require './lib/cell'
 require './lib/ship'
 require './lib/board'
+require './lib/turn'
 
 class Menu
   attr_reader :selection
